Extract Carroucell reader into official plugin spectrochempy-carroucel
- Move read_carroucell from core to plugins/spectrochempy-carroucel/
- Remove carroucell-specific logic from core importer.py (special key mapping)
- Remove carroucell from core filetype registry
- Remove read_carroucell from core lazyimport api_methods
- Remove read_carroucell references from core reader docstrings
- Remove carroucell import from core readers __init__.py
- Move tests to plugin directory with updated scp.carroucel namespace
- Register plugin handlers: importer.infer_filetype_key for .carroucell
