+++ plugins/spectrochempy-carroucel/pyproject.toml
@@ -1,0 +1,23 @@
+[build-system]
+requires = ["setuptools>=64"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "spectrochempy-carroucel"
+version = "0.1.0"
+description = "Carroucell experiment reader for SpectroChemPy"
+requires-python = ">=3.11"
+dependencies = [
+    "spectrochempy>=0.8,<0.9",
+    "xlrd",
+    "scipy",
+]
+
+[project.entry-points."spectrochempy.plugins"]
+carroucel = "spectrochempy_carroucel:CarroucelPlugin"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
