+++ src/spectrochempy/core/readers/filetypes.py
@@ -22,7 +22,6 @@ class FileTypeRegistry:
             ("excel", "Microsoft Excel files (*.xls)"),
             ("zip", "Compressed folder of data files (*.zip)"),
             ("quadera", "Quadera ascii files (*.asc)"),
-            ("carroucell", "Carroucell files (*spa)"),
             ("galactic", "GRAMS/Thermo Galactic files (*.spc)"),
             ("wire", "Renishaw WiRE files (*.wdf)"),
         ]
