+++ src/spectrochempy/core/readers/read_jcamp.py
@@ -120,7 +120,6 @@ def read_jcamp(*paths, **kwargs):
 
     read_csv : Read CSV files (:file:`.csv`).
     read_matlab : Read Matlab files (:file:`.mat`, :file:`.dso`).
-    read_carroucell : Read files in a directory after a carroucell experiment.
     read_wire : Read Renishaw Wire files (:file:`.wdf`).
 
     """
