+++ src/spectrochempy/core/readers/read_matlab.py
@@ -119,7 +119,6 @@ def read_matlab(*paths, **kwargs):
 
     read_csv : Read CSV files (:file:`.csv`).
     read_jcamp : Read Infrared JCAMP-DX files (:file:`.jdx`, :file:`.dx`).
-    read_carroucell : Read files in a directory after a carroucell experiment.
     read_wire : Read Renishaw Wire files (:file:`.wdf`).
 
     Examples
