+++ src/spectrochempy/core/readers/read_spc.py
@@ -120,7 +120,6 @@ def read_spc(*paths, **kwargs):
     read_csv : Read CSV files (:file:`.csv`).
     read_matlab : Read Matlab files (:file:`.mat`, :file:`.dso`).
     read_jcamp : Read Infrared JCAMP-DX files (:file:`.jdx`, :file:`.dx`).
-    read_carroucell : Read files in a directory after a carroucell experiment.
     read_wire : Read Renishaw Wire files (:file:`.wdf`).
 
     Examples
