+++ src/spectrochempy/core/readers/read_wire.py
@@ -886,7 +886,6 @@ def read_wire(*paths, **kwargs):
     read_csv : Read CSV files (:file:`.csv`).
     read_matlab : Read Matlab files (:file:`.mat`, :file:`.dso`).
     read_jcamp : Read Infrared JCAMP-DX files (:file:`.jdx`, :file:`.dx`).
-    read_carroucell : Read files in a directory after a carroucell experiment.
 
     Examples
     --------
